pyproject.toml
[project]
name = "ps01-support-agent"
version = "0.1.0"
description = "A2A Translator agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "click>=8.1.8",
    "google-genai>=1.0.0", 
    "pydantic>=2.11.4",
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
    "python-dotenv>=1.0.1"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"